Opacities default values on float

+++ Sources/JsHelper/Shared/Extensions/FloatExtensions.swift
@@ -1,6 +1,6 @@
 public extension Float {
     // Opacity
-    static let opacityHigh: Double = 0.8
-    static let opacityMedium: Double = 0.5
-    static let opacityLow: Double = 0.35
+    static let opacityHigh: Float = 0.8
+    static let opacityMedium: Float = 0.5
+    static let opacityLow: Float = 0.35
 }
